requirements.txt
asgiref==3.2.3
Django==3.0.2
django-admin-sortable2==0.7.6
django-ckeditor==5.9.0
django-crispy-forms==1.8.1
django-extra-settings==0.1.1
django-filer==1.7.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
easy-thumbnails==2.7
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
six==1.15.0
sqlparse==0.3.0
Unidecode==1.1.1
